requirements.txt
playwright
requests
python-dotenv
beautifulsoup4
python-telegram-bot
